feat: added check role middleware

+++ routes/api.php
@@ -19,11 +19,12 @@ use Illuminate\Support\Facades\Route;
 Route::prefix('auth')->group(function () {
     Route::post('/login', [AuthController::class, 'login']);
     Route::post('/register', [AuthController::class, 'register']);
+    Route::post('/register/admins', [AuthController::class, 'registerAdmin'])->middleware(['auth:sanctum', 'role:root']);
     Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 });
 
 Route::prefix('users')->group(function () {
-    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
+    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware(['auth:sanctum', 'role:user,admin,root']);
 });
 
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
